move prune_unused_indices, freeze_orbitals, and linearize_decoder
