+++ src/openfermion/ops/__init__.py
@@ -13,14 +13,11 @@
 from ._binary_polynomial import BinaryPolynomial
 from ._polynomial_tensor import PolynomialTensor, general_basis_change
 from ._quadratic_hamiltonian import QuadraticHamiltonian
-from ._symbolic_operator import (SymbolicOperator,
-                                 prune_unused_indices)
+from ._symbolic_operator import SymbolicOperator
 
 # Imports out of alphabetical order to avoid circular dependency.
-from ._binary_code import (BinaryCode,
-                           linearize_decoder)
+from ._binary_code import BinaryCode
 from ._fermion_operator import (FermionOperator,
-                                freeze_orbitals,
                                 normal_ordered)
 from ._interaction_operator import InteractionOperator
 from ._qubit_operator import QubitOperator
