+++ src/openfermion/ops/_symbolic_operator_test.py
@@ -19,8 +19,7 @@ import numpy
 from openfermion.config import EQ_TOLERANCE
 from openfermion.utils._testing_utils import EqualsTester
 
-from openfermion.ops._symbolic_operator import (SymbolicOperator,
-                                                prune_unused_indices)
+from openfermion.ops._symbolic_operator import SymbolicOperator
 
 
 class DummyOperator1(SymbolicOperator):
@@ -961,12 +960,6 @@ class SymbolicOperatorTest2(unittest.TestCase):
     def test_tracenorm_zero(self):
         op = SymbolicOperator()
         self.assertFalse(op.induced_norm())
-
-    def test_prune(self):
-        op = DummyOperator1(((1, 1), (8, 1), (3, 0)), 0.5)
-        op = prune_unused_indices(op)
-        expected = DummyOperator1(((0, 1), (2, 1), (1, 0)), 0.5)
-        self.assertTrue(expected == op)
 
 
 class DeprecatedFunctionsTest(unittest.TestCase):
