+++ src/openfermion/transforms/__init__.py
@@ -20,7 +20,8 @@ from ._binary_codes import (bravyi_kitaev_code,
                             parity_code,
                             weight_two_segment_code,
                             weight_one_binary_addressing_code,
-                            weight_one_segment_code)
+                            weight_one_segment_code,
+                            linearize_decoder)
 from ._bravyi_kitaev import bravyi_kitaev
 from ._bravyi_kitaev_tree import bravyi_kitaev_tree
 from ._conversion import (get_fermion_operator,
